+++ sdk/nodejs/defaultPrivileges.ts
@@ -25,7 +25,7 @@ import * as utilities from "./utilities";
  * });
  * ```
  *
- * > This content is derived from https://github.com/terraform-providers/terraform-provider-postgresql/blob/master/website/docs/r/default_privileges.html.markdown.
+ * > This content is derived from https://github.com/terraform-providers/terraform-provider-postgresql/blob/master/website/docs/r/postgresql_default_privileges.html.markdown.
  */
 export class DefaultPrivileges extends pulumi.CustomResource {
     /**